requirements.txt
discord
epicstore_api
python-dotenv